Feature/ranking be

See merge request s04-final/s04p31a306!29

+++ runnershi_be/src/main/java/com/ssafy/runnershi/repository/UserInfoRepository.java
@@ -3,7 +3,7 @@ package com.ssafy.runnershi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.ssafy.runnershi.entity.UserInfo;
 
-public interface UserInfoRepository extends JpaRepository<UserInfo, Integer> {
+public interface UserInfoRepository extends JpaRepository<UserInfo, Long> {
 
   public UserInfo findByUserId_UserId(String userId);
